consumer/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
ImageHash==4.2.0
multidict==5.1.0
numpy==1.20.3
pika==1.2.0
pika-stubs==0.1.3
Pillow==8.2.0
PyWavelets==1.1.1
requests==2.25.1
scipy==1.6.3
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
